Restructuring stuff, showing config path
