+++ ward/tests/test_util.py
@@ -1,8 +1,7 @@
 import os
-import shutil
-import tempfile
 from pathlib import Path
 
+from ward.tests.utilities import make_project
 from ward import test, using, fixture
 from ward.testing import each
 from ward.util import (
@@ -29,22 +28,6 @@ def _():
     project_root = find_project_root([])
     fs_root = os.path.normpath(os.path.abspath(os.sep))
     assert project_root == Path(fs_root)
-
-
-def make_project(root_file: str):
-    tempdir = Path(tempfile.gettempdir())
-    paths = [
-        tempdir / "project/a/b/c",
-        tempdir / "project/a/d",
-        tempdir / "project/a",
-    ]
-    for path in paths:
-        path.mkdir(parents=True, exist_ok=True)
-
-    root_file = tempdir / f"project/{root_file}"
-    with open(root_file, "w+", encoding="utf-8"):
-        yield tempdir / "project"
-    shutil.rmtree(tempdir / "project")
 
 
 @fixture
